# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-tenant CRM SaaS API (Phase 1 foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orgs/{orgId}:
    get:
      operationId: getOrg
      tags: [orgs]
      summary: Org details (requires membership)
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "403":
          description: Not a member
    patch:
      operationId: updateOrg
      tags: [orgs]
      summary: Update org name/plan (owner/admin)
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [orgs]
      summary: Permanently delete an organization (owner only)
      description: Requires the authenticated organization owner to type DELETE. All organization data and private objects are permanently removed; global user accounts are preserved.
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationDeletionConfirmation"
      responses:
        "200":
          description: Organization deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationDeletionResponse"
        "400":
          description: Confirmation must be exactly DELETE
        "401":
          description: Not authenticated
        "403":
          description: Only the organization owner can delete
        "404":
          description: Organization not found
        "409":
          description: A deletion is already in progress
        "502":
          description: An external deletion dependency failed; retry is safe
        "500":
          description: The relational purge failed and was rolled back; retry is safe
  /organizations/{orgId}:
    delete:
      operationId: deleteOrganizationAlias
      tags: [orgs]
      summary: Permanently delete an organization (owner-only alias)
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationDeletionConfirmation"
      responses:
        "200":
          description: Organization deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationDeletionResponse"
        "400":
          description: Confirmation must be exactly DELETE
        "401":
          description: Not authenticated
        "403":
          description: Only the organization owner can delete
        "404":
          description: Organization not found
        "409":
          description: A deletion is already in progress
        "502":
          description: An external deletion dependency failed; retry is safe
        "500":
          description: The relational purge failed and was rolled back; retry is safe
  /orgs/{orgId}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List org members
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: inviteMember
      tags: [members]
      summary: Invite a user by email with a role (admin+)
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInvite"
      responses:
        "201":
          description: Created pre-provisioned membership and attempted invitation delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberInvitationResponse"
  /orgs/{orgId}/members/{memberId}/resend-invite:
    post:
      operationId: resendMemberInvite
      tags: [members]
      summary: Resend an invitation for an existing member (admin+)
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: Existing membership and attempted invitation delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberInvitationResponse"
  /orgs/{orgId}/members/{memberId}:
    patch:
      operationId: updateMemberRole
      tags: [members]
      summary: Change a member's role (admin+)
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRoleUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: removeMember
      tags: [members]
      summary: Remove a member (admin+)
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "204":
          description: Removed
  /orgs/{orgId}/features:
    get:
      operationId: listFeatures
      tags: [features]
      summary: Feature entitlements for the org
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Entitlements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureEntitlement"
    put:
      operationId: updateFeatures
      tags: [features]
      summary: Set the org's enabled features (owner/admin)
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureSelection"
      responses:
        "200":
          description: Updated org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
...
components:
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
    MemberId:
      name: memberId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        clerkId: { type: string }
        email: { type: string }
        fullName: { type: ["string", "null"] }
...
    Organization:
      type: object
      required: [id, name, slug, plan, enabledFeatures]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        plan:
          type: string
          enum: [essential, professional, enterprise, custom]
        enabledFeatures:
          type: array
          items: { type: string }
        stripeCustomerId: { type: ["string", "null"] }
        stripeSubscriptionId: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
    OrganizationUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        plan:
          type: string
          enum: [essential, professional, enterprise, custom]
    OrganizationDeletionConfirmation:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation: { type: string, enum: [DELETE] }
    OrganizationDeletionResponse:
      type: object
      additionalProperties: false
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
...
    Member:
      type: object
      required: [id, role, user]
      properties:
        id: { type: string }
        role:
          type: string
          enum: [owner, admin, manager, user, viewer]
        user:
          $ref: "#/components/schemas/User"
        createdAt: { type: ["string", "null"] }
    MemberInvite:
      type: object
      required: [email, role]
      properties:
        email: { type: string }
        fullName: { type: string }
        role:
          type: string
          enum: [admin, manager, user, viewer]
    InvitationDelivery:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [sent, failed]
        message: { type: ["string", "null"] }
    MemberInvitationResponse:
      allOf:
        - $ref: "#/components/schemas/Member"
        - type: object
          required: [delivery]
          properties:
            delivery:
              $ref: "#/components/schemas/InvitationDelivery"
...
    MemberRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [owner, admin, manager, user, viewer]
    FeatureEntitlement:
      type: object
      required: [featureKey, enabled]
      properties:
        featureKey: { type: string }
        enabled: { type: boolean }
    FeatureSelection:
      type: object
      required: [features]
      properties:
        features:
          type: array
          items: { type: string }
...