# lib/db/src/schema/platform.ts
import {
  boolean,
  jsonb,
  numeric,
  pgTable,
  text,
  timestamp,
  uuid,
} from "drizzle-orm/pg-core";
import { organizations } from "./organizations";
import { users } from "./users";

export const featureEntitlements = pgTable("feature_entitlements", {
  id: uuid("id").primaryKey().defaultRandom(),
  orgId: uuid("org_id")
    .notNull()
    .references(() => organizations.id, { onDelete: "cascade" }),
  featureKey: text("feature_key").notNull(),
  enabled: boolean("enabled").default(true).notNull(),
  featureConfig: jsonb("feature_config").default({}),
  createdAt: timestamp("created_at", { withTimezone: true }).defaultNow().notNull(),
});

export const usageLogs = pgTable("usage_logs", {
  id: uuid("id").primaryKey().defaultRandom(),
  orgId: uuid("org_id")
    .notNull()
    .references(() => organizations.id, { onDelete: "cascade" }),
  // Usage is organization telemetry; retain it while detaching the deleted
  // actor rather than cascading a surviving organization's history.
  userId: uuid("user_id").references(() => users.id, { onDelete: "set null" }),
  featureKey: text("feature_key").notNull(),
  action: text("action").notNull(),
  metadata: jsonb("metadata").default({}),
  createdAt: timestamp("created_at", { withTimezone: true }).defaultNow().notNull(),
});

export type FeatureEntitlement = typeof featureEntitlements.$inferSelect;
export type UsageLog = typeof usageLogs.$inferSelect;